Make dockup an umbrella project
