+++ apps/dockup/.gitignore
@@ -1,0 +1,9 @@
+/_build
+/cover
+/deps
+erl_crash.dump
+*.ez
+/.vagrant
+/nginx_config_dir
+/workdir
+/projects.dets
