+++ apps/dockup/lib/dockup/command.ex
