+++ apps/dockup/lib/dockup/fake_command.ex
