+++ apps/dockup/lib/dockup/shell_command.ex
