+++ apps/dockup/priv/adjectives.txt
